# docs/openapi.json
{"openapi": "3.0.2", "info": {"title": "Unofficial SureFlap API", "description": "SureFlap API is a simple, yet powerful RESTful API for products from [Sure Petcare](https://www.surepetcare.com).", "version": "4.1.1"}, "paths": {"/dashboard": {"get": {"tags": ["Dashboard"], "summary": "Get Dashboard", "operationId": "get_dashboard_dashboard_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Dashboard"}}}}}}}, "/devices": {"get": {"tags": ["Device"], "summary": "Get Devices", "operationId": "get_devices_devices_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Devices Devices Get", "type": "array", "items": {"anyOf": [{"$ref": "#/components/schemas/HubShort"}, {"$ref": "#/components/schemas/FlapShort"}]}}}}}}}}, "/devices/{device_id}": {"get": {"tags": ["Device"], "summary": "Get Device By Id", "operationId": "get_device_by_id_devices__device_id__get", "parameters": [{"required": true, "schema": {"title": "Device Id", "type": "integer"}, "name": "device_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Device By Id Devices  Device Id  Get", "anyOf": [{"$ref": "#/components/schemas/Hub"}, {"$ref": "#/components/schemas/Flap"}]}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households": {"get": {"tags": ["Household"], "summary": "Get Households", "operationId": "get_households_households_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Households Households Get", "type": "array", "items": {"$ref": "#/components/schemas/HouseholdShort"}}}}}}}}, "/households/{household_id}": {"get": {"tags": ["Household"], "summary": "Get Household By Id", "operationId": "get_household_by_id_households__household_id__get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Household"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households/{household_id}/pets": {"get": {"tags": ["Pet"], "summary": "Get Pets From Household", "operationId": "get_pets_from_household_households__household_id__pets_get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Pets From Household Households  Household Id  Pets Get", "type": "array", "items": {"$ref": "#/components/schemas/PetShort"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households/{household_id}/pets/{pet_id}/location": {"get": {"tags": ["Pet"], "summary": "Get Pet Location", "operationId": "get_pet_location_households__household_id__pets__pet_id__location_get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}, {"required": true, "schema": {"title": "Pet Id", "type": "integer"}, "name": "pet_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PetLocation"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Pet"], "summary": "Set Pet Location", "operationId": "set_pet_location_households__household_id__pets__pet_id__location_post", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}, {"required": true, "schema": {"title": "Pet Id", "type": "integer"}, "name": "pet_id", "in": "path"}], "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/PetLocationSet"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PetLocationUpdate"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households/{household_id}/pets/location": {"get": {"tags": ["Pet"], "summary": "Get Pets Locations", "operationId": "get_pets_locations_households__household_id__pets_location_get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Pets Locations Households  Household Id  Pets Location Get", "type": "array", "items": {"$ref": "#/components/schemas/PetLocations"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households/{household_id}/pets/{pet_id}": {"get": {"tags": ["Pet"], "summary": "Get Pet", "operationId": "get_pet_households__household_id__pets__pet_id__get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}, {"required": true, "schema": {"title": "Pet Id", "type": "integer"}, "name": "pet_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Pet"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households/{household_id}/users": {"get": {"tags": ["User"], "summary": "Get Users From Household", "operationId": "get_users_from_household_households__household_id__users_get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"title": "Response Get Users From Household Households  Household Id  Users Get", "type": "array", "items": {"$ref": "#/components/schemas/UserShort"}}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/households/{household_id}/users/{user_id}": {"get": {"tags": ["User"], "summary": "Get User", "operationId": "get_user_households__household_id__users__user_id__get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}, {"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/User"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "deprecated": true}}, "/households/{household_id}/users/{user_id}/photo": {"get": {"tags": ["User"], "summary": "Get User Photo", "operationId": "get_user_photo_households__household_id__users__user_id__photo_get", "parameters": [{"required": true, "schema": {"title": "Household Id", "type": "integer"}, "name": "household_id", "in": "path"}, {"required": true, "schema": {"title": "User Id", "type": "integer"}, "name": "user_id", "in": "path"}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Photo"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "deprecated": true}}}, "components": {"schemas": {"Control": {"title": "Control", "type": "object", "properties": {"led_mode": {"title": "Led Mode", "type": "number"}, "pairing_mode": {"title": "Pairing Mode", "type": "number"}}}, "Curfew": {"title": "Curfew", "type": "object", "properties": {"enabled": {"title": "Enabled", "type": "boolean"}, "lock_time": {"title": "Lock Time", "type": "string", "format": "time"}}}, "CurfewLock": {"title": "CurfewLock", "type": "object", "properties": {"delay_time": {"title": "Delay Time", "type": "number"}, "lock_time": {"title": "Lock Time", "type": "string", "format": "time"}, "locked": {"title": "Locked", "type": "boolean"}, "permissions": {"title": "Permissions", "type": "number"}, "unlock_time": {"title": "Unlock Time", "type": "string", "format": "time"}}}, "Dashboard": {"title": "Dashboard", "type": "object", "properties": {"devices": {"title": "Devices", "type": "array", "items": {"anyOf": [{"$ref": "#/components/schemas/HubShort"}, {"$ref": "#/components/schemas/FlapShort"}]}}, "households": {"title": "Households", "type": "array", "items": {"$ref": "#/components/schemas/HouseholdDashboard"}}, "pets": {"title": "Pets", "type": "array", "items": {"$ref": "#/components/schemas/Pet"}}, "photos": {"title": "Photos", "type": "array", "items": {"$ref": "#/components/schemas/Photo"}}, "tags": {"title": "Tags", "type": "array", "items": {"$ref": "#/components/schemas/Tag"}}, "user": {"$ref": "#/components/schemas/User"}}}, "Device": {"title": "Device", "type": "object", "properties": {"firmware": {"title": "Firmware", "type": "number"}, "hardware": {"title": "Hardware", "type": "number"}}}, "Direction": {"title": "Direction", "enum": [1, 2], "description": "An enumeration."}, "Flap": {"title": "Flap", "type": "object", "properties": {"control": {"$ref": "#/components/schemas/FlapControl"}, "created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "household_id": {"title": "Household Id", "type": "number"}, "id": {"title": "Id", "type": "number"}, "index": {"title": "Index", "type": "number"}, "mac_adress": {"title": "Mac Adress", "type": "string"}, "name": {"title": "Name", "type": "string"}, "pairing_at": {"title": "Pairing At", "type": "string", "format": "date-time"}, "parent": {"$ref": "#/components/schemas/HubShort"}, "parent_device_id": {"title": "Parent Device Id", "type": "number"}, "product_id": {"title": "Product Id", "type": "number"}, "status": {"$ref": "#/components/schemas/FlapStatus"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "FlapControl": {"title": "FlapControl", "type": "object", "properties": {"curfew": {"$ref": "#/components/schemas/Curfew"}, "fast_polling": {"title": "Fast Polling", "type": "boolean"}}}, "FlapShort": {"title": "FlapShort", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "household_id": {"title": "Household Id", "type": "number"}, "id": {"title": "Id", "type": "number"}, "index": {"title": "Index", "type": "number"}, "name": {"title": "Name", "type": "string"}, "pairing_at": {"title": "Pairing At", "type": "string", "format": "date-time"}, "parent_device_id": {"title": "Parent Device Id", "type": "number"}, "product_id": {"title": "Product Id", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "FlapStatus": {"title": "FlapStatus", "type": "object", "properties": {"battery": {"title": "Battery", "type": "number"}, "learn_mode": {"title": "Learn Mode", "type": "boolean"}, "locking": {"$ref": "#/components/schemas/Locking"}, "online": {"title": "Online", "type": "boolean"}, "signal": {"$ref": "#/components/schemas/Signal"}, "version": {"$ref": "#/components/schemas/FlapVersion"}}}, "FlapVersion": {"title": "FlapVersion", "type": "object", "properties": {"lcd": {"$ref": "#/components/schemas/Device"}, "rf": {"$ref": "#/components/schemas/Device"}}}, "HTTPValidationError": {"title": "HTTPValidationError", "type": "object", "properties": {"detail": {"title": "Detail", "type": "array", "items": {"$ref": "#/components/schemas/ValidationError"}}}}, "Household": {"title": "Household", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "id": {"title": "Id", "type": "number"}, "name": {"title": "Name", "type": "string"}, "pets": {"title": "Pets", "type": "array", "items": {"$ref": "#/components/schemas/PetShort"}}, "share_code": {"title": "Share Code", "type": "string"}, "timezone_id": {"title": "Timezone Id", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "users": {"title": "Users", "type": "array", "items": {"$ref": "#/components/schemas/UserShort"}}, "version": {"title": "Version", "type": "string"}}}, "HouseholdDashboard": {"title": "HouseholdDashboard", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "id": {"title": "Id", "type": "number"}, "invites": {"title": "Invites", "type": "array", "items": {"$ref": "#/components/schemas/Invite"}}, "name": {"title": "Name", "type": "string"}, "share_code": {"title": "Share Code", "type": "string"}, "timezone_id": {"title": "Timezone Id", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "users": {"title": "Users", "type": "array", "items": {"$ref": "#/components/schemas/UserShort"}}, "version": {"title": "Version", "type": "string"}}}, "HouseholdShort": {"title": "HouseholdShort", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "id": {"title": "Id", "type": "number"}, "name": {"title": "Name", "type": "string"}, "timezone_id": {"title": "Timezone Id", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "Hub": {"title": "Hub", "type": "object", "properties": {"children": {"title": "Children", "type": "array", "items": {"$ref": "#/components/schemas/FlapShort"}}, "control": {"$ref": "#/components/schemas/Control"}, "created_at": {"title": "Created At", "type": "string"}, "household_id": {"title": "Household Id", "type": "number"}, "mac_adress": {"title": "Mac Adress", "type": "string"}, "name": {"title": "Name", "type": "string"}, "product_id": {"title": "Product Id", "type": "number"}, "serial_number": {"title": "Serial Number", "type": "string"}, "status": {"$ref": "#/components/schemas/HubStatus"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "HubShort": {"title": "HubShort", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "household_id": {"title": "Household Id", "type": "number"}, "id": {"title": "Id", "type": "number"}, "name": {"title": "Name", "type": "string"}, "product_id": {"title": "Product Id", "type": "number"}, "serial_number": {"title": "Serial Number", "type": "string"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "HubStatus": {"title": "HubStatus", "type": "object", "properties": {"led_mode": {"title": "Led Mode", "type": "string"}, "online": {"title": "Online", "type": "boolean"}, "pairing_mode": {"title": "Pairing Mode", "type": "number"}, "version": {"$ref": "#/components/schemas/HubVersion"}}}, "HubVersion": {"title": "HubVersion", "type": "object", "properties": {"device": {"$ref": "#/components/schemas/Device"}}}, "Invite": {"title": "Invite", "type": "object", "properties": {"acceptor_user_id": {"title": "Acceptor User Id", "type": "number"}, "code": {"title": "Code", "type": "string"}, "created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "creator_user_id": {"title": "Creator User Id", "type": "number"}, "email_adress": {"title": "Email Adress", "type": "string"}, "id": {"title": "Id", "type": "number"}, "owner": {"title": "Owner", "type": "boolean"}, "status": {"title": "Status", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "user": {"$ref": "#/components/schemas/InviteRoles"}, "version": {"title": "Version", "type": "string"}}}, "InviteRoles": {"title": "InviteRoles", "type": "object", "properties": {"acceptor": {"$ref": "#/components/schemas/InviteUser"}, "creator": {"$ref": "#/components/schemas/InviteUser"}}}, "InviteUser": {"title": "InviteUser", "type": "object", "properties": {"id": {"title": "Id", "type": "number"}, "name": {"title": "Name", "type": "string"}, "photo": {"$ref": "#/components/schemas/Photo"}, "photo_id": {"title": "Photo Id", "type": "number"}}}, "Locking": {"title": "Locking", "type": "object", "properties": {"curfew": {"$ref": "#/components/schemas/CurfewLock"}, "mode": {"title": "Mode", "type": "number"}}}, "Notifications": {"title": "Notifications", "type": "object", "properties": {"animal_movement": {"title": "Animal Movement", "type": "boolean"}, "curfew": {"title": "Curfew", "type": "boolean"}, "device_status": {"title": "Device Status", "type": "boolean"}, "feeding_activity": {"title": "Feeding Activity", "type": "boolean"}, "household_management": {"title": "Household Management", "type": "boolean"}, "intruder_movements": {"title": "Intruder Movements", "type": "boolean"}, "low_battery": {"title": "Low Battery", "type": "boolean"}, "new_device_pet": {"title": "New Device Pet", "type": "boolean"}, "photos": {"title": "Photos", "type": "boolean"}}}, "Pet": {"title": "Pet", "type": "object", "properties": {"comments": {"title": "Comments", "type": "string"}, "created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "gender": {"title": "Gender", "type": "number"}, "household_id": {"title": "Household Id", "type": "number"}, "id": {"title": "Id", "type": "number"}, "name": {"title": "Name", "type": "string"}, "photo": {"$ref": "#/components/schemas/Photo"}, "photo_id": {"title": "Photo Id", "type": "number"}, "position": {"$ref": "#/components/schemas/Position"}, "species_id": {"title": "Species Id", "type": "number"}, "tag_id": {"title": "Tag Id", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "PetLocation": {"title": "PetLocation", "type": "object", "properties": {"location": {"title": "Location", "type": "string"}, "pet_id": {"title": "Pet Id", "type": "number"}}}, "PetLocationSet": {"title": "PetLocationSet", "required": ["where"], "type": "object", "properties": {"where": {"$ref": "#/components/schemas/Direction"}}}, "PetLocationUpdate": {"title": "PetLocationUpdate", "type": "object", "properties": {"pet_id": {"title": "Pet Id", "type": "number"}, "tag_id": {"title": "Tag Id", "type": "number"}, "user_id": {"title": "User Id", "type": "number"}, "where": {"title": "Where", "type": "number"}, "since": {"title": "Since", "type": "string", "format": "date-time"}}}, "PetLocations": {"title": "PetLocations", "type": "object", "properties": {"duration": {"title": "Duration", "type": "string"}, "location": {"title": "Location", "type": "string"}, "name": {"title": "Name", "type": "string"}, "since": {"title": "Since", "type": "string", "format": "date-time"}}}, "PetShort": {"title": "PetShort", "type": "object", "properties": {"comments": {"title": "Comments", "type": "string"}, "created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "gender": {"title": "Gender", "type": "number"}, "household_id": {"title": "Household Id", "type": "number"}, "id": {"title": "Id", "type": "number"}, "name": {"title": "Name", "type": "string"}, "photo_id": {"title": "Photo Id", "type": "number"}, "species_id": {"title": "Species Id", "type": "number"}, "tag_id": {"title": "Tag Id", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "Photo": {"title": "Photo", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "id": {"title": "Id", "type": "number"}, "location": {"title": "Location", "type": "string"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "uploading_user_id": {"title": "Uploading User Id", "type": "number"}, "version": {"title": "Version", "type": "string"}}}, "Position": {"title": "Position", "type": "object", "properties": {"device_id": {"title": "Device Id", "type": "number"}, "since": {"title": "Since", "type": "string", "format": "date-time"}, "tag_id": {"title": "Tag Id", "type": "number"}, "where": {"title": "Where", "type": "number"}}}, "Signal": {"title": "Signal", "type": "object", "properties": {"device_rssi": {"title": "Device Rssi", "type": "string"}, "hub_rssi": {"title": "Hub Rssi", "type": "string"}}}, "Tag": {"title": "Tag", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "id": {"title": "Id", "type": "number"}, "supported_product_ids": {"title": "Supported Product Ids", "type": "array", "items": {"type": "number"}}, "tag": {"title": "Tag", "type": "string"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}}}, "User": {"title": "User", "type": "object", "properties": {"country_id": {"title": "Country Id", "type": "number"}, "created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "email_adress": {"title": "Email Adress", "type": "string"}, "first_name": {"title": "First Name", "type": "string"}, "id": {"title": "Id", "type": "number"}, "language_id": {"title": "Language Id", "type": "number"}, "last_name": {"title": "Last Name", "type": "string"}, "marketing_opt_in": {"title": "Marketing Opt In", "type": "boolean"}, "notifications": {"$ref": "#/components/schemas/Notifications"}, "terms_accepted": {"title": "Terms Accepted", "type": "string", "format": "date-time"}, "time_format": {"title": "Time Format", "type": "number"}, "updated_at": {"title": "Updated At", "type": "string"}, "version": {"title": "Version", "type": "string"}, "weight_units": {"title": "Weight Units", "type": "number"}}}, "UserShort": {"title": "UserShort", "type": "object", "properties": {"created_at": {"title": "Created At", "type": "string", "format": "date-time"}, "id": {"title": "Id", "type": "number"}, "owner": {"title": "Owner", "type": "boolean"}, "updated_at": {"title": "Updated At", "type": "string", "format": "date-time"}, "version": {"title": "Version", "type": "string"}, "write": {"title": "Write", "type": "boolean"}}}, "ValidationError": {"title": "ValidationError", "required": ["loc", "msg", "type"], "type": "object", "properties": {"loc": {"title": "Location", "type": "array", "items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}}, "msg": {"title": "Message", "type": "string"}, "type": {"title": "Error Type", "type": "string"}}}}}}